Split sessions directory into sessions/core and sesssions/http
