+++ packages/core/src/sessions/core/constants.ts
@@ -1,0 +1,5 @@
+// Expiration timeouts in seconds
+export const SESSION_DEFAULT_INACTIVITY_TIMEOUT = 15 * 60; // 15 minutes
+export const SESSION_DEFAULT_ABSOLUTE_TIMEOUT = 7 * 24 * 60 * 60; // 1 week
+
+export const SESSION_DEFAULT_GARBAGE_COLLECTOR_PERIODICITY = 50;
