+++ packages/core/src/sessions/core/create-session.ts
@@ -1,4 +1,4 @@
-import { generateToken } from '../common';
+import { generateToken } from '../../common';
 import { Session } from './session';
 import { SessionState } from './session-state.interface';
 import { SessionStore } from './session-store';
