+++ packages/core/src/sessions/http/fetch-user.interface.ts
@@ -1,3 +1,3 @@
-import { ServiceManager } from '../core';
+import { ServiceManager } from '../../core';
 
 export type FetchUser = (id: string|number, services: ServiceManager) => Promise<any>;
