+++ packages/core/src/sessions/http/index.ts
@@ -1,0 +1,2 @@
+export { UseSessions } from './use-sessions.hook';
+export { FetchUser } from './fetch-user.interface';
