+++ packages/core/src/sessions/http/set-session-cookie.spec.ts
@@ -1,5 +1,8 @@
-import { strictEqual } from 'assert';
-import { Config, HttpResponse, HttpResponseOK } from '../core';
+// std
+import { deepStrictEqual, strictEqual } from 'assert';
+
+// FoalTS
+import { Config, HttpResponse, HttpResponseOK } from '../../core';
 import {
   SESSION_DEFAULT_COOKIE_HTTP_ONLY,
   SESSION_DEFAULT_COOKIE_NAME,
@@ -8,23 +11,45 @@ import {
   SESSION_DEFAULT_SAME_SITE_ON_CSRF_ENABLED,
   SESSION_USER_COOKIE_NAME,
 } from './constants';
-import { removeSessionCookie } from './remove-session-cookie';
+import { Session } from '../core';
+import { setSessionCookie } from './set-session-cookie';
 
-describe('removeSessionCookie', () => {
+describe('setSessionCookie', () => {
 
+  let token: string;
+  let csrfToken: string;
+  let session: Session;
   let response: HttpResponse;
 
-  beforeEach(() => response = new HttpResponseOK());
+  beforeEach(() => {
+    response = new HttpResponseOK();
+    token = 'xxx';
+    csrfToken = 'yyy',
+    session = new Session(
+      {} as any,
+      {
+        content: {
+          csrfToken,
+        },
+        createdAt: Math.trunc(Date.now() / 1000),
+        flash: {},
+        id: token,
+        updatedAt: Math.trunc(Date.now() / 1000),
+        userId: null,
+      },
+      { exists: true },
+    );
+  });
 
   describe('should set a session cookie in the response', () => {
 
     context('given no configuration option is provided', () => {
 
-      beforeEach(() => removeSessionCookie(response));
+      beforeEach(() => setSessionCookie(response, session));
 
       it('with the proper default name and value.', () => {
         const { value } = response.getCookie(SESSION_DEFAULT_COOKIE_NAME);
-        strictEqual(value, '');
+        strictEqual(value, token);
       });
 
       it('with the proper default "domain" directive.', () => {
@@ -52,9 +77,9 @@ describe('removeSessionCookie', () => {
         strictEqual(options.secure, undefined);
       });
 
-      it('with the proper "maxAge" directive.', () => {
+      it('with the proper "expires" directive.', () => {
         const { options } = response.getCookie(SESSION_DEFAULT_COOKIE_NAME);
-        strictEqual(options.maxAge, 0);
+        deepStrictEqual(options.expires, new Date(session.expirationTime * 1000));
       });
 
     });
@@ -70,7 +95,7 @@ describe('removeSessionCookie', () => {
         Config.set('settings.session.cookie.path', '/foo');
         Config.set('settings.session.cookie.sameSite', 'strict');
         Config.set('settings.session.cookie.secure', true);
-        removeSessionCookie(response);
+        setSessionCookie(response, session);
       });
 
       afterEach(() => {
@@ -84,7 +109,7 @@ describe('removeSessionCookie', () => {
 
       it('with the proper default name and value.', () => {
         const { value } = response.getCookie(cookieName);
-        strictEqual(value, '');
+        strictEqual(value, token);
       });
 
       it('with the proper default "domain" directive.', () => {
@@ -112,9 +137,9 @@ describe('removeSessionCookie', () => {
         strictEqual(options.secure, true);
       });
 
-      it('with the proper "maxAge" directive.', () => {
+      it('with the proper "expires" directive.', () => {
         const { options } = response.getCookie(cookieName);
-        strictEqual(options.maxAge, 0);
+        deepStrictEqual(options.expires, new Date(session.expirationTime * 1000));
       });
 
     });
@@ -131,7 +156,7 @@ describe('removeSessionCookie', () => {
 
       context('given no configuration option is provided', () => {
 
-        beforeEach(() => removeSessionCookie(response));
+        beforeEach(() => setSessionCookie(response, session));
 
         it('with the proper default "sameSite" directive.', () => {
           const { options } = response.getCookie(SESSION_DEFAULT_COOKIE_NAME);
@@ -144,10 +169,12 @@ describe('removeSessionCookie', () => {
 
         beforeEach(() => {
           Config.set('settings.session.cookie.sameSite', 'strict');
-          removeSessionCookie(response);
+          setSessionCookie(response, session);
         });
 
-        afterEach(() => Config.remove('settings.session.cookie.sameSite'));
+        afterEach(() => {
+          Config.remove('settings.session.cookie.sameSite');
+        });
 
         it('with the proper default "sameSite" directive.', () => {
           const { options } = response.getCookie(SESSION_DEFAULT_COOKIE_NAME);
@@ -162,11 +189,11 @@ describe('removeSessionCookie', () => {
 
       context('given no configuration option is provided', () => {
 
-        beforeEach(() => removeSessionCookie(response));
+        beforeEach(() => setSessionCookie(response, session));
 
         it('with the proper default name and value.', () => {
           const { value } = response.getCookie(SESSION_DEFAULT_CSRF_COOKIE_NAME);
-          strictEqual(value, '');
+          strictEqual(value, csrfToken);
         });
 
         it('with the proper default "domain" directive.', () => {
@@ -194,9 +221,9 @@ describe('removeSessionCookie', () => {
           strictEqual(options.secure, undefined);
         });
 
-        it('with the proper "maxAge" directive.', () => {
+        it('with the proper "expires" directive.', () => {
           const { options } = response.getCookie(SESSION_DEFAULT_CSRF_COOKIE_NAME);
-          strictEqual(options.maxAge, 0);
+          deepStrictEqual(options.expires, new Date(session.expirationTime * 1000));
         });
 
       });
@@ -211,8 +238,8 @@ describe('removeSessionCookie', () => {
           Config.set('settings.session.cookie.httpOnly', true);
           Config.set('settings.session.cookie.path', '/foo');
           Config.set('settings.session.cookie.sameSite', 'strict');
-          Config.set('settings.session.cookie.secure', 'true');
-          removeSessionCookie(response);
+          Config.set('settings.session.cookie.secure', true);
+          setSessionCookie(response, session);
         });
 
         afterEach(() => {
@@ -226,7 +253,7 @@ describe('removeSessionCookie', () => {
 
         it('with the proper default name and value.', () => {
           const { value } = response.getCookie(csrfCookieName);
-          strictEqual(value, '');
+          strictEqual(value, csrfToken);
         });
 
         it('with the proper default "domain" directive.', () => {
@@ -254,9 +281,9 @@ describe('removeSessionCookie', () => {
           strictEqual(options.secure, true);
         });
 
-        it('with the proper "maxAge" directive.', () => {
+        it('with the proper "expires" directive.', () => {
           const { options } = response.getCookie(csrfCookieName);
-          strictEqual(options.maxAge, 0);
+          deepStrictEqual(options.expires, new Date(session.expirationTime * 1000));
         });
 
       });
@@ -267,7 +294,7 @@ describe('removeSessionCookie', () => {
 
   context('given the CSRF protection is disabled in the config', () => {
 
-    beforeEach(() => removeSessionCookie(response));
+    beforeEach(() => setSessionCookie(response, session));
 
     it('should not set a CSRF cookie in the response.', () => {
       const { value } = response.getCookie(SESSION_DEFAULT_CSRF_COOKIE_NAME);
@@ -276,17 +303,19 @@ describe('removeSessionCookie', () => {
 
   });
 
-  context('given the "user" argument is true', () => {
+  context('given the "user" argument is defined', () => {
 
     describe('should set a "user" cookie in the response', () => {
 
+      const user = 'foo';
+
       context('given no configuration option is provided', () => {
 
-        beforeEach(() => removeSessionCookie(response, true));
+        beforeEach(() => setSessionCookie(response, session, user));
 
         it('with the proper default name and value.', () => {
           const { value } = response.getCookie(SESSION_USER_COOKIE_NAME);
-          strictEqual(value, '');
+          strictEqual(value, user);
         });
 
         it('with the proper default "domain" directive.', () => {
@@ -315,9 +344,9 @@ describe('removeSessionCookie', () => {
           strictEqual(options.secure, undefined);
         });
 
-        it('with the proper "maxAge" directive.', () => {
+        it('with the proper "expires" directive.', () => {
           const { options } = response.getCookie(SESSION_USER_COOKIE_NAME);
-          strictEqual(options.maxAge, 0);
+          deepStrictEqual(options.expires, new Date(session.expirationTime * 1000));
         });
 
       });
@@ -329,8 +358,8 @@ describe('removeSessionCookie', () => {
           Config.set('settings.session.cookie.httpOnly', true);
           Config.set('settings.session.cookie.path', '/foo');
           Config.set('settings.session.cookie.sameSite', 'strict');
-          Config.set('settings.session.cookie.secure', 'true');
-          removeSessionCookie(response, true);
+          Config.set('settings.session.cookie.secure', true);
+          setSessionCookie(response, session, user);
         });
 
         afterEach(() => {
@@ -343,7 +372,7 @@ describe('removeSessionCookie', () => {
 
         it('with the proper default name and value.', () => {
           const { value } = response.getCookie(SESSION_USER_COOKIE_NAME);
-          strictEqual(value, '');
+          strictEqual(value, user);
         });
 
         it('with the proper default "domain" directive.', () => {
@@ -372,9 +401,9 @@ describe('removeSessionCookie', () => {
           strictEqual(options.secure, true);
         });
 
-        it('with the proper "maxAge" directive.', () => {
+        it('with the proper "expires" directive.', () => {
           const { options } = response.getCookie(SESSION_USER_COOKIE_NAME);
-          strictEqual(options.maxAge, 0);
+          deepStrictEqual(options.expires, new Date(session.expirationTime * 1000));
         });
 
       });
@@ -383,11 +412,11 @@ describe('removeSessionCookie', () => {
 
   });
 
-  context('given the "user" argument is false or undefined', () => {
+  context('given the "user" argument is empty or not defined', () => {
 
-    beforeEach(() => removeSessionCookie(response));
+    beforeEach(() => setSessionCookie(response, session));
 
-    it('should not set a "user" cookie in the response.', () => {
+    it('should not set a CSRF cookie in the response.', () => {
       const { value } = response.getCookie(SESSION_USER_COOKIE_NAME);
       strictEqual(value, undefined);
     });
