+++ packages/core/src/sessions/http/set-session-cookie.ts
@@ -1,5 +1,5 @@
 // FoalTS
-import { Config, CookieOptions, HttpResponse } from '../core';
+import { Config, CookieOptions, HttpResponse } from '../../core';
 import {
   SESSION_DEFAULT_COOKIE_HTTP_ONLY,
   SESSION_DEFAULT_COOKIE_NAME,
@@ -8,19 +8,21 @@ import {
   SESSION_DEFAULT_SAME_SITE_ON_CSRF_ENABLED,
   SESSION_USER_COOKIE_NAME,
 } from './constants';
+import { Session } from '../core';
 
 /**
- * Deletes the browser cookie containing the session token.
+ * Sends the session token in a cookie.
  *
- * If the CSRF protection is enabled, it also deletes the CSRF cookie containing the CSRF token.
+ * If the CSRF protection is enabled, it also sends the CSRF token in a CSRF cookie.
  *
- * If the `user` argument is true, it also deletes the "user" cookie.
+ * If a "user" argument is provided, it also sends its value in a "user" cookie.
  *
  * @export
- * @param {HttpResponse} response - The HTTP response
- * @param {boolean} [user] - Specify if the "user" cookie should be deleted.
+ * @param {HttpResponse} response - The HTTP response.
+ * @param {Session} session - The session object.
+ * @param {string} [user] - The content of the "user" cookie if any.
  */
-export function removeSessionCookie(response: HttpResponse, user?: boolean): void {
+export function setSessionCookie(response: HttpResponse, session: Session, user?: string): void {
   const cookieName = Config.get('settings.session.cookie.name', 'string', SESSION_DEFAULT_COOKIE_NAME);
 
   const csrfEnabled = Config.get('settings.session.csrf.enabled', 'boolean', false);
@@ -31,27 +33,27 @@ export function removeSessionCookie(response: HttpResponse, user?: boolean): voi
 
   const options: CookieOptions = {
     domain: Config.get('settings.session.cookie.domain', 'string'),
-    maxAge: 0,
+    expires: new Date(session.expirationTime * 1000),
     path: Config.get('settings.session.cookie.path', 'string', SESSION_DEFAULT_COOKIE_PATH),
     sameSite,
-    secure: Config.get('settings.session.cookie.secure', 'boolean'),
+    secure: Config.get('settings.session.cookie.secure', 'boolean')
   };
 
-  response.setCookie(cookieName, '', {
+  response.setCookie(cookieName, session.getToken(), {
     ...options,
     httpOnly: Config.get('settings.session.cookie.httpOnly', 'boolean', SESSION_DEFAULT_COOKIE_HTTP_ONLY),
   });
 
   if (csrfEnabled) {
     const csrfCookieName = Config.get('settings.session.csrf.cookie.name', 'string', SESSION_DEFAULT_CSRF_COOKIE_NAME);
-    response.setCookie(csrfCookieName, '', {
+    response.setCookie(csrfCookieName, session.get<string|undefined>('csrfToken') || '', {
       ...options,
       httpOnly: false,
     });
   }
 
   if (user) {
-    response.setCookie(SESSION_USER_COOKIE_NAME, '', {
+    response.setCookie(SESSION_USER_COOKIE_NAME, user, {
       ...options,
       httpOnly: false,
     });
