+++ packages/core/src/sessions/index.ts
@@ -1,7 +1,10 @@
-export { createSession } from './create-session';
-export { FetchUser } from './fetch-user.interface';
-export { readSession } from './read-session';
-export { Store, SessionStore, SessionAlreadyExists } from './session-store';
-export { Session } from './session';
-export { SessionState } from './session-state.interface';
-export { UseSessions } from './use-sessions.hook';
+export {
+  createSession,
+  readSession,
+  Store,
+  SessionStore,
+  SessionAlreadyExists,
+  Session,
+  SessionState,
+} from './core';
+export { UseSessions, FetchUser } from './http';
